css cleanup, refactor code
