+++ client/app/styles/queue/_evaluate_decision.scss
@@ -1,3 +1,11 @@
+.case-timeline {
+  display: flex;
+
+  .case-type {
+    width: 15%;
+  }
+}
+
 .grayLineStyling {
   width: 5px;
   background: $color-gray-dark;
